gltfio: introduce TextureProvider API.
This allows clients to provide their own asychronous texture decoders
for various mime-typed images. This is similar to MaterialProvider in
that it's a plug-in component for gltfio.

There are two motivations for this: to decouple gltfio from STB and
to make it easier to integrate support for BasisU textures.

This also has the side effect of simplifying ResourceLoader, since the
texture decoding jobs have been moved out.